imageconfig: fstab configuration
Allow controlling if `/etc/fstab` is written directly from image config
instead of through an experimental environment variable only.

This introduces an additional argument to the generation of filesystem
configuration stages and allows for there to be *no* configuration
stages for either mount units or the fstab.

With this change it is possible to define an image type that uses the
DSP [1] fully needing no state about the filesystem stored inside the
operating system tree.

There's a lot of mechanical work in the commit as we don't really do
defaults on our image configs and I don't want to invert the option.
Hence we explicitly enable the writing of fstab on all stages that
don't write mount units.

Most importantly, no manifests change here.

[1]: https://uapi-group.org/specifications/specs/discoverable_partitions_specification/

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>