many: use fstab config for `mount.usr`
`GenImageKernelOptions` determines if the `mount.usr` argument needs to
be added based on if there's an `fstab` file being written; it was using
the mount units configuration for this but this no longer controls if
the `fstab` is written.

Check the `fstab` option instead to determine if `mount.usr` is
necessary (other bits still apply, e.g. if there's a separate `/usr`
mount point at all).

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>